server/requirements.txt
audioread==3.0.1
av==12.1.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
ctranslate2==4.3.1
cycler==0.12.1
decorator==5.1.1
dnspython==2.6.1
eventlet==0.36.1
faster-whisper==1.0.2
filelock==3.15.3
Flask==3.0.3
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.0
googletrans==4.0.0rc1
greenlet==3.0.3
gTTS==2.5.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.6.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.23.4
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keyboard==0.13.5
kiwisolver==1.4.5
lazy_loader==0.4
libretranslatepy==2.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.0
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
noisereduce==3.0.2
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.18.0
openai-whisper==20231117
packaging==24.1
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.27.1
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rfc3986==1.5.0
scikit-learn==1.5.0
scipy==1.13.1
setuptools==70.3.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
SpeechRecognition==3.10.4
sympy==1.12.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
translate==3.6.1
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wsproto==1.2.0
